pull out shared method
