+++ src/connectors/common.ts
@@ -13,3 +13,9 @@ export function getQsParam(name: string) {
 
     return decodeURIComponent(results[2].replace(/\+/g, ' '));
 }
+
+export function b64Decode(str: string) {
+    return decodeURIComponent(Array.prototype.map.call(atob(str), (c: string) => {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+    }).join(''));
+}
